# perf: use zap's Check() to prevent useless allocs