Zap provides [an optional API](https://pkg.go.dev/go.uber.org/zap#example-Logger.Check) to prevent allocating fields that will never be used because the configured log level is higher than the log emitted.

We had good results using this trick [in Mercure](https://github.com/dunglas/mercure/pull/554) and more recently [in FrankenPHP](https://github.com/dunglas/frankenphp/pull/1005). This is also the trick that has been used in https://github.com/caddyserver/caddy/pull/6541.

This is especially important for `DEBUG` and `INFO` level logs (which are usually disabled in production), as well as for benchmarks (where logs are usually entirely disabled).

As you can see in the added benchmark, this significantly improves performance:

Before (no logs):

```
goos: darwin
goarch: arm64
pkg: github.com/caddyserver/caddy/v2/modules/caddyhttp
cpu: Apple M1 Pro
BenchmarkServer_LogRequest_NopLogger
BenchmarkServer_LogRequest_NopLogger-10    	 2797832	       433.9 ns/op	     408 B/op	       3 allocs/op
PASS
ok  	github.com/caddyserver/caddy/v2/modules/caddyhttp	2.150s
```

After (no logs):

```
goos: darwin
goarch: arm64
pkg: github.com/caddyserver/caddy/v2/modules/caddyhttp
cpu: Apple M1 Pro
BenchmarkServer_LogRequest_NopLogger
BenchmarkServer_LogRequest_NopLogger-10    	55410920	        21.63 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/caddyserver/caddy/v2/modules/caddyhttp	2.693s
```

Our FrankenPHP benchmarks show that the server can handle ~1% more requests per second by applying this patch (a Hello World page). In some cases, the performance gain can be more important (streaming for instance, which logs many things at the `DEBUG` level).

The optimized syntax is a bit more verbose, but not that much IMHO.