@francislavoie 2024-10-06
I think this isn't working correctly. I have `debug` enabled, but `reverse_proxy`'s `upstream roundtrip` logs aren't showing up anymore.