@francislavoie 2024-10-06
Oh sorry I'm an absolute dummy, the `reverse_proxy` was just not running (directive order etc) :joy: it's working fine.