@jum 2024-10-17
I am seeing the use of zap.Error() in cases where the previous line contains a check for zap.Warnlevel in a few of these changes, which surely looks like a mistake to me, as it changes the level of the log line to error where previously a warning was used. At least the issuing of error messages for writing errors is back, the pull request https://github.com/caddyserver/caddy/pull/6532 does not have any effect any longer.