@jum 2024-10-18
I am not that familiar with zap logging, but looking at the source code of zap it should not happen at all. The Check function records the level that is being checked and uses it on ce.Write, but still I do get these "aborting with incomplete response" messages on ErrorLevel. Anyone any clue how that is possible?