@jum 2024-10-18
Ok, ignore these comments from me above. I had a small error in my CI and I build accidentally against "latest" instead of "master", I checked my docker container and it has 2.8.4 as the version, way before the patches. Sorry.